+++ platform/lib/patchbay/forum/types/jev_kind.ex
@@ -1,0 +1,3 @@
+defmodule Patchbay.Forum.Types.JevKind do
+  use Ash.Type.Enum, values: [:tool_defect, :usage_question, :setup_problem, :other]
+end